// src/app/api/model/dokumentStatus.ts
/**
 * Next-Gen App API
 * Die REST Web-Schnittstelle zu der Next-Gen App Anwendung
 *
 * The version of the OpenAPI document: v1
 * Contact: info@brunnerinformatik.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type DokumentStatus = 0 | 10 | 20 | 40 | 50 | 70 | 71 | 100 | 105 | 107 | 110 | 120;

export const DokumentStatus = {
    Undefiniert: 0 as DokumentStatus,
    Importiert: 10 as DokumentStatus,
    InArbeit: 20 as DokumentStatus,
    TeilSigniert: 40 as DokumentStatus,
    SigniertGesperrt: 50 as DokumentStatus,
    Gesendet: 70 as DokumentStatus,
    ErhaltBestaetigt: 71 as DokumentStatus,
    Bewilligt: 100 as DokumentStatus,
    BewilligtMassnahmen: 105 as DokumentStatus,
    AntwortErhalten: 107 as DokumentStatus,
    ErledigtOhneAntwort: 110 as DokumentStatus,
    Abgelehnt: 120 as DokumentStatus
};